# src/cmd/http/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server",
        "title": "API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Loggearse dentro de la aplicación de NeoHome",
                "parameters": [
                    {
                        "description": "Password y username",
                        "name": "authDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.LoginResponse"
                        }
                    },
                    "403": {
                        "description": "Credenciales inválidas",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "La sesión no existe. Probablemente porque la eliminaron",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/auth/password": {
            "patch": {
                "tags": [
                    "auth"
                ],
                "summary": "Actualizar la contraseña",
                "parameters": [
                    {
                        "description": "newPassword",
                        "name": "authDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAuthPasswordDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Credenciales inválidas",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "La sesión no existe. Probablemente porque la eliminaron",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Refrescar sesión",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de refresco, es decir, de sesión",
                        "name": "X-Refresh",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "No es un token válido JWT",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "403": {
                        "description": "No está el token de refresco en el header X-Refresh",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "El token no tiene un usuario registrado en la BD",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "La sesión no existe. Probablemente porque la eliminaron",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Registrase dentro de la aplicación de AlpacaTatto",
                "parameters": [
                    {
                        "description": "name, username, email, password, role",
                        "name": "authDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.LoginResponse"
                        }
                    },
                    "403": {
                        "description": "Credenciales inválidas",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "La sesión no existe. Probablemente porque la eliminaron",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
...
        "/api/generators/code": {
            "post": {
                "tags": [
                    "code"
                ],
                "summary": "Crear codigo",
                "parameters": [
                    {
                        "description": "usesRemaining, type y duration",
                        "name": "newCodeDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NewCodeDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/generators/code/verify/{code}": {
            "get": {
                "tags": [
                    "code"
                ],
                "summary": "Verificar si un codigo es valido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "type",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Error con la base de datos",
                        "schema": {
                            "$ref": "#/definitions/utils.ProblemDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.LoginResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                },
                "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
...
        "dto.AuthDto": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.NewCodeDTO": {
            "type": "object",
            "required": [
                "duration",
                "type",
                "usesRemaining"
            ],
            "properties": {
                "duration": {
                    "type": "integer",
                    "minimum": 1
                },
                "idUser": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "usesRemaining": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "dto.RegisterDto": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "role",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
...
        "dto.UpdateAuthPasswordDTO": {
            "type": "object",
            "required": [
                "newPassword"
            ],
            "properties": {
                "newPassword": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
...
        "model.Role": {
            "type": "string",
            "enum": [
                "user",
                "tattooArtist",
                "admin"
            ],
            "x-enum-varnames": [
                "USER_ROLE",
                "TATTOO_ARTIST_ROLE",
                "ADMIN_ROLE"
            ]
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Role"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "utils.ErrorProblemDetails": {
            "type": "object",
            "properties": {
                "param": {
                    "type": "string",
                    "example": "3"
                },
                "pointer": {
                    "type": "string",
                    "example": "max"
                },
                "title": {
                    "type": "string",
                    "example": "name"
                }
            }
        },
        "utils.ProblemDetails": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Detalle técnico del error"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.ErrorProblemDetails"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Descripción del problema para mostrar al usuario"
                },
                "type": {
                    "type": "string",
                    "example": "/docs/errors/errorPointer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "BearerJWTToken in Authorization Header",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}